refactor to custom ReactRouterSidebarItem
